refactor: move `CenterModal`
